Limit clothoid length parameter to 10km to avoid rendering blowup
When start and end point of the clothoid are close to each other and the
tangent angles are colinear, the solution can become a gigantic circle.
This blows up the rendering of the associated road in Blender.
Therefore, do not return a new when the length becomes to large and
instead keep the solution with the last valid parameters. The user is
notified that no new valid solution has been found via a Blender
statusbar report.