Implement creation of roads with multiple geometry sections
For this we convert the two point modal operator into a multi point
modal operator and with that add geometries to the road for each
consecutive point. We also introduce holding the SHIFT key to add
additional geomtry sections. Changing the start heading is moved to the
ALT key.

Note that the curve type for the different geometry section must be
identical for each road for now. Where road in this context means the
OpenDRIVE notion of road.