Rewrite and improve the junction connection road tool
Make the tool create roads for each individual connection. For this a
rewrite of the lane width calculation and user interface is necessary.
Fix some bugs in the width calculation along the way.
Also add lane information to junction joints for snapping and decouple
connecting road snapping and incoming road snapping such that junction
connecting roads can be created independent of the incoming roads.